\name{convert}
\alias{convert}
\title{Converts raw DMS extract variables to appropriate data types.}
\usage{
  convert(renamed)
}
\arguments{
  \item{renamed}{The output of the \code{\link(rename)}
  function.}
}
\description{
  This function converts variables from the DMS extracts
  into appropriate R data types; currently, that means
  converting character dates and datetimes to Dates and
  POSIXct.
}
\details{
  Currently, this function just converts character dates
  and datetimes to Date and POSIXct where appropriate. This
  function expects the variable names generated by the
  \code{\link(rename)} function, and is only meant to be
  used internally by \code{\link(clean_to1)}
}

