\docType{data}
\name{country_codes}
\alias{country_codes}
\title{Countries by name and their ISO 3166-1 alpha-3 codes.}
\source{
  \url{http://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}
}
\description{
  A dataset containing the names and three-letter ISO
  country codes of 249 countries, dependent territories,
  and special areas of geographical interest.
}
\details{
  \itemize{ \item name. Name of country/area.  \item code
  ISO 3166-1 alpha-3 code for the country/area.  ... }
}
\keyword{datasets}

