\name{rename}
\alias{rename}
\title{Renames variables in the raw DMS extract to more manageable strings.}
\usage{
rename(originals)
}
\arguments{
  \item{originals}{The list of data frames generated
  interally by \code{\link{clean_to1}}}
}
\description{
This function renames the raw DMS extracts to be more
consistent and easier to type.
}
\details{
This function is only meant to be used internally by
\code{\link{clean_to1}}.
}

