\name{to1check-package}
\alias{to1check-package}
\alias{to1check}
\docType{package}
\title{
Reporting and quality checks for TBESC TO 1 data
}
\description{
This package generates a variety of reports and data quality checks from TBESC TO 1 data extracts.
}
\details{
\tabular{ll}{
Package: \tab to1check\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-03-28\cr
License: \tab What license is it under?\cr
}
An overview of how to use the package, including the most important functions
}
\author{
Matt Parker

Maintainer: Matt Parker <mattmparker@gmail.com>
}
\keyword{ package }
\seealso{
}
\examples{
}
