\name{to1check-package}
\alias{to1check-package}
\alias{to1check}
\docType{package}
\title{
Reporting and quality checks for TBESC TO 1 data
}
\description{
This package generates a variety of reports and data quality checks from TBESC TO 1 data extracts.
}
\details{
\tabular{ll}{
Package: \tab to1check\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-04-10\cr
License: \tab MIT\cr
}
This package makes it easier to run reports and data quality checks against the
TO 1 DMS extracts. Its key function is \code{\link{clean_to1}}, which converts
the raw DMS extracts into a clean, prepared list of data.frames. The remainder
of its functions generate various reports and plots.
}
\author{
Matt Parker

Maintainer: Matt Parker <mattmparker@gmail.com>
}
\keyword{ package }
\seealso{
}
\examples{
}
